+++ _studio/mfx_lib/decode/h264/src/mfx_h264_dec_decode.cpp
@@ -1654,7 +1654,7 @@ mfxStatus VideoDECODEH264::DecodeFrame(mfxFrameSurface1 *surface_out, UMC::H264D
     pFrame->setWasDisplayed();
 
     TRACE_EVENT(MFX_TRACE_API_AVC_DISPLAYINFO_TASK, EVENT_TYPE_INFO, TR_KEY_DECODE_BASIC_INFO, make_event_data(
-        pFrame->m_PicOrderCnt[0], pFrame->m_PicOrderCnt[1], pFrame->FrameNum(), pFrame->wasDisplayed(), pFrame->wasOutputted()));
+        pFrame->m_PicOrderCnt[0], pFrame->m_PicOrderCnt[1], pFrame->FrameNum(), (uint32_t)pFrame->wasDisplayed(), (uint32_t)pFrame->wasOutputted()));
 
     return sts;
 }
